requirements.txt
astroid==2.0.4
autopep8==1.4.3
colorama==0.4.0
dj-database-url==0.5.0
Django==2.2.2
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==0.0.7
django-crispy-forms==1.7.2
django-heroku==0.3.1
django-widget-tweaks==1.4.3
djangorestframework==3.9.4
gunicorn==19.9.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
notification==0.2.1
Pillow==5.3.0
psycopg2==2.7.7
psycopg2-binary==2.7.7
pycodestyle==2.4.0
pylint==2.1.1
pytz==2018.7
six==1.11.0
typed-ast==1.1.0
whitenoise==4.1.2
wrapt==1.10.11
